Almost done Pondeuse

+++ include/Carapace.h
@@ -7,13 +7,13 @@
 class Carapace : public Accessoire {
   private:
 
-    float protection;
-    float ralentissement;
+    double protection;
+    double ralentissement;
 
 
   public:
     Carapace(const Carapace & c);
-    Carapace(float protection, float ralentissement);
+    Carapace(double protection, double ralentissement);
     ~Carapace();
 
 };
